+++ common/src/client/java/io/github/jamalam360/sort_it_out/client/ClientSortableContainer.java
@@ -5,17 +5,28 @@ import io.github.jamalam360.sort_it_out.sort.SortableContainer;
 import net.minecraft.client.Minecraft;
 import net.minecraft.world.Container;
 import net.minecraft.world.inventory.AbstractContainerMenu;
+import net.minecraft.world.inventory.Slot;
 import net.minecraft.world.item.ItemStack;
 
 import java.util.Arrays;
+import java.util.HashMap;
 
 public class ClientSortableContainer implements SortableContainer {
 	private final int size;
+	private final Container container;
 	private final AbstractContainerMenu menu;
+	private final HashMap<Integer, Integer> containerToMenuSlots = new HashMap<>();
 
 	public ClientSortableContainer(Container container) {
 		this.size = container.getContainerSize();
+		this.container = container;
 		this.menu = Minecraft.getInstance().player.containerMenu;
+
+		for (Slot slot : this.menu.slots) {
+			if (slot.container == this.container) {
+				containerToMenuSlots.put(slot.getContainerSlot(), slot.index);
+			}
+		}
 	}
 
 	@Override
@@ -25,7 +36,7 @@ public class ClientSortableContainer implements SortableContainer {
 
 	@Override
 	public ItemStack getItem(int i) {
-		return this.menu.getSlot(i).getItem();
+		return this.container.getItem(i);
 	}
 
 	@Override
@@ -79,23 +90,23 @@ public class ClientSortableContainer implements SortableContainer {
 	}
 
 	private void pickup(int slot) {
-		ClientPacketWorkQueue.INSTANCE.submit(new ClientPacketWorkQueue.PickupItemAction(this.menu, slot, this.getItem(slot).copy()));
-		this.menu.getSlot(slot).set(ItemStack.EMPTY);
+		ClientPacketWorkQueue.INSTANCE.submit(new ClientPacketWorkQueue.PickupItemAction(this.menu, containerToMenuSlots.getOrDefault(slot, slot), this.getItem(slot).copy()));
+		this.container.setItem(slot, ItemStack.EMPTY);
 	}
 
 	private void place(int slot, ItemStack item) {
-		ClientPacketWorkQueue.INSTANCE.submit(new ClientPacketWorkQueue.PlaceItem(this.menu, slot, item, ItemStack.EMPTY));
-		this.menu.getSlot(slot).set(item);
+		ClientPacketWorkQueue.INSTANCE.submit(new ClientPacketWorkQueue.PlaceItem(this.menu, containerToMenuSlots.getOrDefault(slot, slot), item, ItemStack.EMPTY));
+		this.container.setItem(slot, item);
 	}
 
 	private void placeAndPickupItemInSlot(int slot, ItemStack carried) {
-		ClientPacketWorkQueue.INSTANCE.submit(new ClientPacketWorkQueue.PlaceItem(this.menu, slot, carried, this.getItem(slot).copy()));
-		this.menu.getSlot(slot).set(carried);
+		ClientPacketWorkQueue.INSTANCE.submit(new ClientPacketWorkQueue.PlaceItem(this.menu, containerToMenuSlots.getOrDefault(slot, slot), carried, this.getItem(slot).copy()));
+		this.container.setItem(slot, carried);
 	}
 
 	private void placeExpectingRemainder(int slot, ItemStack newSlotContents, ItemStack expectedRemainder) {
-		ClientPacketWorkQueue.INSTANCE.submit(new ClientPacketWorkQueue.PlaceItem(this.menu, slot, newSlotContents, expectedRemainder));
-		this.menu.getSlot(slot).set(newSlotContents);
+		ClientPacketWorkQueue.INSTANCE.submit(new ClientPacketWorkQueue.PlaceItem(this.menu, containerToMenuSlots.getOrDefault(slot, slot), newSlotContents, expectedRemainder));
+		this.container.setItem(slot, newSlotContents);
 	}
 
 	private int findWorkingSlot(ItemStack workingItem, int... blacklist) {
@@ -105,7 +116,7 @@ public class ClientSortableContainer implements SortableContainer {
 				continue;
 			}
 
-			if (this.menu.getSlot(i).getItem().isEmpty() || this.menu.getSlot(i).getItem().getItem() != workingItem.getItem()) {
+			if (this.getItem(i).isEmpty() || this.getItem(i).getItem() != workingItem.getItem()) {
 				return i;
 			}
 		}